refactor: extract SessionRotationService behind an explicit RotationHost
Phase 3 of the engine.ts decomposition (#867). The managed session-file
rotation family (startup/runtime auto-rotate, transcript rewrite,
raw-context compaction outside the fresh tail, rotate-with-backup) moves
to src/session-rotation.ts (~1,500 lines). Rotation orchestrates
reconciliation, compaction, and summarizer resolution, so it runs against
a RotationHost interface that explicitly inventories that coupling; the
engine passes a host object and keeps thin public wrappers for the four
externally-used entry points (plugin commands and startup hook).

Shared rotate types (TranscriptRewriteRequest/Replacement,
ContextEngineMaintenanceResult, RotateSessionStorage*Result) move with
the service and are re-exported from engine.ts for existing importers;
TranscriptReconcileResult moves to reconcile-plan.ts to keep the module
graph acyclic. The two near-identical database-backup blocks now share
createRotateDatabaseBackup(), with call-site-specific failure reporting
kept apart deliberately.

Full suite: 1279 tests passing.
